Move Type{Load,Save} helper traits into crate-wide utils
